split up apply tests
