+++ packages/apply/test/apply.test.ts
@@ -1,6 +1,5 @@
-import { apply, changesConsideringParents, validateApply } from "../src";
+import { apply } from "../src";
 import { loadFile } from "@covector/files";
-import { run } from "effection";
 import { it } from "@effection/jest";
 import mockConsole, { RestoreConsole } from "jest-mock-console";
 import fixtures from "fixturez";
@@ -601,567 +600,7 @@ describe("package file apply bump (snapshot)", () => {
   });
 });
 
-describe("list changes considering parents", () => {
-  let restoreConsole: RestoreConsole;
-  beforeEach(() => {
-    restoreConsole = mockConsole(["log", "dir"]);
-  });
-  afterEach(() => {
-    restoreConsole();
-  });
-
-  it("adds changes for dependency", function* () {
-    const assembledChanges = {
-      releases: {
-        all: {
-          dependencies: undefined,
-          manager: "javascript",
-          path: undefined,
-          pkg: "all",
-          type: "minor",
-        },
-      },
-    };
-
-    const config = {
-      packages: {
-        "yarn-workspace-base-pkg-a": {
-          path: "./packages/pkg-a/",
-          manager: "javascript",
-          dependencies: ["yarn-workspace-base-pkg-b", "all"],
-        },
-        "yarn-workspace-base-pkg-b": {
-          path: "./packages/pkg-b/",
-          manager: "javascript",
-          dependencies: ["all"],
-        },
-        all: { version: true },
-      },
-    };
-
-    //@ts-ignore
-    const changes = changesConsideringParents({ assembledChanges, config });
-
-    expect({
-      //@ts-ignore
-      consoleLog: console.log.mock.calls,
-      //@ts-ignore
-      consoleDir: console.dir.mock.calls,
-      changes,
-    }).toMatchSnapshot();
-  });
-
-  it("bumps patch due to dependency bump", function* () {
-    const assembledChanges = {
-      releases: {
-        "yarn-workspace-base-pkg-a": {
-          dependencies: undefined,
-          manager: "javascript",
-          path: undefined,
-          pkg: "all",
-          type: "patch",
-        },
-        all: {
-          dependencies: undefined,
-          manager: "javascript",
-          path: undefined,
-          pkg: "all",
-          type: "minor",
-        },
-      },
-    };
-
-    const config = {
-      packages: {
-        "yarn-workspace-base-pkg-a": {
-          path: "./packages/pkg-a/",
-          manager: "javascript",
-          dependencies: ["yarn-workspace-base-pkg-b", "all"],
-        },
-        "yarn-workspace-base-pkg-b": {
-          path: "./packages/pkg-b/",
-          manager: "javascript",
-          dependencies: ["all"],
-        },
-        all: { version: true },
-      },
-    };
-
-    //@ts-ignore
-    const changes = changesConsideringParents({ assembledChanges, config });
-
-    expect({
-      //@ts-ignore
-      consoleLog: console.log.mock.calls,
-      //@ts-ignore
-      consoleDir: console.dir.mock.calls,
-      changes,
-    }).toMatchSnapshot();
-  });
-
-  it("rolls up the parent bumps", function* () {
-    const changesFiles = [
-      {
-        releases: {
-          "pkg-c": "patch",
-          "pkg-d": "major",
-        },
-        summary:
-          "This should patch bump up the pkg-[number] line to where pkg-one also receives a bump. The pkg-c writes to a patch so we patch bump up that line too.",
-        meta: {},
-      },
-      {
-        releases: {
-          "pkg-b": "minor",
-        },
-        summary:
-          "The pkg-b doesn't have a dep bump, but can dep bump pkg-a and pkg-overall with a patch.",
-        meta: {},
-      },
-    ];
-
-    const assembledChanges = {
-      releases: {
-        "pkg-b": {
-          type: "minor",
-          changes: [changesFiles[1]],
-        },
-        "pkg-c": {
-          type: "patch",
-          changes: [changesFiles[0]],
-        },
-        "pkg-d": {
-          type: "major",
-          changes: [changesFiles[0]],
-        },
-      },
-    };
-
-    const config = {
-      packages: {
-        "pkg-overall": {
-          path: "./packages/pkg-overall/",
-          manager: "javascript",
-          dependencies: ["pkg-a", "pkg-b"],
-        },
-        "pkg-a": {
-          path: "./packages/pkg-a/",
-          manager: "javascript",
-          dependencies: ["pkg-c"],
-        },
-        "pkg-b": {
-          path: "./packages/pkg-b/",
-          manager: "javascript",
-          dependencies: [],
-        },
-        "pkg-c": {
-          path: "./packages/pkg-c/",
-          manager: "javascript",
-          dependencies: ["pkg-d"],
-        },
-        "pkg-d": {
-          path: "./packages/pkg-d/",
-          manager: "javascript",
-          dependencies: [],
-        },
-        "pkg-one": {
-          path: "./packages/pkg-one/",
-          manager: "javascript",
-          dependencies: ["pkg-two"],
-        },
-        "pkg-two": {
-          path: "./packages/pkg-two/",
-          manager: "javascript",
-          dependencies: ["pkg-three"],
-        },
-        "pkg-three": {
-          path: "./packages/pkg-three/",
-          manager: "javascript",
-          dependencies: ["pkg-four"],
-        },
-        "pkg-four": {
-          path: "./packages/pkg-four/",
-          manager: "javascript",
-          dependencies: ["pkg-d"],
-        },
-      },
-    };
-
-    //@ts-ignore
-    const changes = changesConsideringParents({ assembledChanges, config });
-    // console.error(changes)
-
-    // these are directly defined in the change files
-    expect(changes.releases["pkg-b"].type).toBe("minor");
-    expect(changes.releases["pkg-c"].type).toBe("patch");
-    expect(changes.releases["pkg-d"].type).toBe("major");
-
-    // these are the top level rolled up
-    expect(changes.releases["pkg-overall"].type).toBe("patch");
-    expect(changes.releases["pkg-a"].type).toBe("patch");
-
-    // rolls up from pkg-d
-    expect(changes.releases["pkg-four"].type).toBe("patch");
-    expect(changes.releases["pkg-three"].type).toBe("patch");
-    expect(changes.releases["pkg-two"].type).toBe("patch");
-    expect(changes.releases["pkg-one"].type).toBe("patch");
-
-    expect({
-      //@ts-ignore
-      consoleLog: console.log.mock.calls,
-      //@ts-ignore
-      consoleDir: console.dir.mock.calls,
-      changes,
-    }).toMatchSnapshot();
-  });
-});
-
-describe("package file apply bump (async/await)", () => {
-  it("bumps single js json", function* () {
-    const jsonFolder = f.copy("pkg.js-single-json");
-
-    const commands = [
-      {
-        dependencies: undefined,
-        manager: "javascript",
-        path: "./",
-        pkg: "js-single-json-fixture",
-        type: "minor",
-        parents: [],
-      },
-    ];
-
-    const config = {
-      packages: {
-        "js-single-json-fixture": {
-          path: "./",
-          manager: "javascript",
-        },
-      },
-    };
-
-    expect.assertions(1);
-    expect(
-      async () =>
-        //@ts-ignore
-        await run(validateApply({ commands, config, cwd: jsonFolder }))
-    ).not.toThrow();
-  });
-
-  it("bumps single rust toml", function* () {
-    const rustFolder = f.copy("pkg.rust-single");
-
-    const commands = [
-      {
-        dependencies: undefined,
-        manager: "rust",
-        path: "./",
-        pkg: "rust-single-fixture",
-        type: "minor",
-        parents: [],
-      },
-    ];
-
-    const config = {
-      packages: {
-        "rust-single-fixture": {
-          path: "./",
-          manager: "rust",
-        },
-      },
-    };
-
-    expect.assertions(1);
-    expect(
-      //@ts-ignore
-      async () => await validateApply({ commands, config, cwd: rustFolder })
-    ).not.toThrow();
-  });
-
-  it("bumps multi js json", function* () {
-    const jsonFolder = f.copy("pkg.js-yarn-workspace");
-
-    const commands = [
-      {
-        dependencies: ["yarn-workspace-base-pkg-b", "all"],
-        manager: "javascript",
-        path: "./",
-        pkg: "yarn-workspace-base-pkg-a",
-        type: "minor",
-        parents: [],
-      },
-      {
-        dependencies: undefined,
-        manager: "javascript",
-        path: undefined,
-        pkg: "yarn-workspace-base-pkg-b",
-        type: "minor",
-        parents: ["yarn-workspace-base-pkg-a"],
-      },
-      {
-        dependencies: undefined,
-        manager: "javascript",
-        path: undefined,
-        pkg: "all",
-        type: "minor",
-        parents: ["yarn-workspace-base-pkg-a", "yarn-workspace-base-pkg-b"],
-      },
-    ];
-
-    const config = {
-      packages: {
-        "yarn-workspace-base-pkg-a": {
-          path: "./packages/pkg-a/",
-          manager: "javascript",
-          dependencies: ["yarn-workspace-base-pkg-b", "all"],
-        },
-        "yarn-workspace-base-pkg-b": {
-          path: "./packages/pkg-b/",
-          manager: "javascript",
-          dependencies: ["all"],
-        },
-        all: { version: true },
-      },
-    };
-
-    expect.assertions(1);
-    expect(
-      //@ts-ignore
-      async () => await validateApply({ commands, config, cwd: jsonFolder })
-    ).not.toThrow();
-  });
-
-  it("bumps multi rust toml", function* () {
-    const rustFolder = f.copy("pkg.rust-multi");
-
-    const commands = [
-      {
-        dependencies: ["rust_pkg_b_fixture"],
-        manager: "rust",
-        path: "./pkg-a/",
-        pkg: "rust_pkg_a_fixture",
-        type: "minor",
-        parents: [],
-      },
-      {
-        dependencies: undefined,
-        manager: "rust",
-        path: "./pkg-b/",
-        pkg: "rust_pkg_b_fixture",
-        type: "minor",
-        parents: [],
-      },
-    ];
-
-    const config = {
-      packages: {
-        rust_pkg_a_fixture: {
-          path: "./pkg-a/",
-          manager: "rust",
-        },
-        rust_pkg_b_fixture: {
-          path: "./pkg-b/",
-          manager: "rust",
-        },
-      },
-    };
-
-    expect.assertions(1);
-    expect(
-      //@ts-ignore
-      async () => await validateApply({ commands, config, cwd: rustFolder })
-    ).not.toThrow();
-  });
-
-  it("bumps multi rust toml with object dep", function* () {
-    const rustFolder = f.copy("pkg.rust-multi-object-dep");
-
-    const commands = [
-      {
-        dependencies: ["rust_pkg_b_fixture"],
-        manager: "rust",
-        path: "./pkg-a/",
-        pkg: "rust_pkg_a_fixture",
-        type: "minor",
-        parents: [],
-      },
-      {
-        dependencies: undefined,
-        manager: "rust",
-        path: "./pkg-b/",
-        pkg: "rust_pkg_b_fixture",
-        type: "minor",
-        parents: [],
-      },
-    ];
-
-    const config = {
-      packages: {
-        rust_pkg_a_fixture: {
-          path: "./pkg-a/",
-          manager: "rust",
-        },
-        rust_pkg_b_fixture: {
-          path: "./pkg-b/",
-          manager: "rust",
-        },
-      },
-    };
-
-    expect.assertions(1);
-    expect(
-      //@ts-ignore
-      async () => await validateApply({ commands, config, cwd: rustFolder })
-    ).not.toThrow();
-  });
-
-  it("bumps multi rust toml with dep missing patch", function* () {
-    const rustFolder = f.copy("pkg.rust-multi-no-patch-dep");
-
-    const commands = [
-      {
-        dependencies: ["rust_pkg_b_fixture"],
-        manager: "rust",
-        path: "./pkg-a/",
-        pkg: "rust_pkg_a_fixture",
-        type: "minor",
-        parents: [],
-      },
-      {
-        dependencies: undefined,
-        manager: "rust",
-        path: "./pkg-b/",
-        pkg: "rust_pkg_b_fixture",
-        type: "minor",
-        parents: [],
-      },
-    ];
-
-    const config = {
-      packages: {
-        rust_pkg_a_fixture: {
-          path: "./pkg-a/",
-          manager: "rust",
-        },
-        rust_pkg_b_fixture: {
-          path: "./pkg-b/",
-          manager: "rust",
-        },
-      },
-    };
-
-    expect.assertions(1);
-    expect(
-      //@ts-ignore
-      async () => await validateApply({ commands, config, cwd: rustFolder })
-    ).not.toThrow();
-  });
-
-  it("bumps multi rust toml as patch with object dep missing patch", function* () {
-    const rustFolder = f.copy("pkg.rust-multi-object-no-patch-dep");
-
-    const commands = [
-      {
-        dependencies: ["rust_pkg_b_fixture"],
-        manager: "rust",
-        path: "./pkg-a/",
-        pkg: "rust_pkg_a_fixture",
-        type: "patch",
-        parents: [],
-      },
-      {
-        dependencies: undefined,
-        manager: "rust",
-        path: "./pkg-b/",
-        pkg: "rust_pkg_b_fixture",
-        type: "patch",
-        parents: [],
-      },
-    ];
-
-    const config = {
-      ...configDefaults,
-      packages: {
-        rust_pkg_a_fixture: {
-          path: "./pkg-a/",
-          manager: "rust",
-        },
-        rust_pkg_b_fixture: {
-          path: "./pkg-b/",
-          manager: "rust",
-        },
-      },
-    };
-
-    expect.assertions(1);
-    //@ts-ignore
-    const validated = yield validateApply({
-      //@ts-ignore
-      commands,
-      config,
-      cwd: rustFolder,
-    });
-    expect(validated).toBe(true);
-  });
-
-  it("bumps multi rust toml as minor with object dep without version number", function* () {
-    let restoreConsole = mockConsole(["error"]);
-
-    const rustFolder = f.copy("pkg.rust-multi-object-path-dep-only");
-
-    const commands = [
-      {
-        dependencies: ["rust_pkg_b_fixture"],
-        manager: "rust",
-        path: "./pkg-a/",
-        pkg: "rust_pkg_a_fixture",
-        type: "minor",
-        parents: [],
-      },
-      {
-        dependencies: undefined,
-        manager: "rust",
-        path: "./pkg-b/",
-        pkg: "rust_pkg_b_fixture",
-        type: "minor",
-        parents: [],
-      },
-    ];
-
-    const config = {
-      packages: {
-        rust_pkg_a_fixture: {
-          path: "./pkg-a/",
-          manager: "rust",
-        },
-        rust_pkg_b_fixture: {
-          path: "./pkg-b/",
-          manager: "rust",
-        },
-      },
-    };
-
-    expect.assertions(2);
-    try {
-      //@ts-ignore
-      yield validateApply({ commands, config, cwd: rustFolder });
-    } catch (e: any) {
-      expect(e.message).toMatch(
-        "rust_pkg_a_fixture has a dependency on rust_pkg_b_fixture, and rust_pkg_b_fixture does not have a version number. " +
-          "This cannot be published. Please pin it to a MAJOR.MINOR.PATCH reference."
-      );
-    }
-
-    expect({
-      //@ts-ignore
-      consoleError: console.error.mock.calls,
-    }).toMatchSnapshot();
-
-    restoreConsole();
-  });
-});
-
-describe("packge file applies preview bump", () => {
+describe("package file applies preview bump", () => {
   let restoreConsole: RestoreConsole;
   beforeEach(() => {
     restoreConsole = mockConsole(["log", "dir"]);
